Added Rock and second minecart script
Added a rock asset to the first puzzle scene, and a second script for the minecart to be used for the first puzzle. The idea is that there'll be a different script for each puzzle. In that way, if we need to instantiate each section by itself, we can switch them out, and if not, we'll write a 'super-script' for the minecart at the same time as we put all the scenes together into one 'super-scene'.