Puzzle scene finally done?
This puzzle scene is now fully complete, I believe.

All textures and meshes are there.
All the scripts work as intended.
I'm afraid that this pull request is going to be huge for whoever to go through, since there are more than just the puzzle scene here since everyone's had their hand in adding stuff to this branch.